requirements.txt
openai>=1.0.0
sentence-transformers>=2.0.0
datasets>=2.0.0
pandas>=1.5.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pyyaml>=6.0
numpy>=1.21.0
python-dotenv>=1.0.0
umap-learn>=0.5.0